+++ frontend/src/logic/ai3.ts
@@ -6,12 +6,9 @@ import 'dotenv/config'
 import { createReactAgent } from '@langchain/langgraph/prebuilt'
 import { AzureChatOpenAI } from '@langchain/openai'
 import { AIMessage, HumanMessage, SystemMessage } from '@langchain/core/messages'
-import { Client } from '@modelcontextprotocol/sdk/client/index.js'
-import { SSEClientTransport } from '@modelcontextprotocol/sdk/client/sse.js'
-import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
-import { loadMcpTools } from '@langchain/mcp-adapters'
 import { sendMessage } from '../pages/api/sse'
 import { mcpServers } from './get-servers.ts'
+import { getTools } from './get-tools.ts'
 
 
 export type Message = {
@@ -38,62 +35,8 @@ export const getLlmResponse = async (messages: Message[], selectedServers: strin
   const selectedMcpServers = mcpServers.filter((server: {name: string}) => selectedServers.includes(server.name));
 
   // Get mcpTools for all servers
-  let mcpTools = [];
-  
-  for (const mcpServer of selectedMcpServers) {
-    const serverHeaders: any = {};
-    if (mcpServer.accessToken) {
-      serverHeaders['x-external-access-token'] = mcpServer.accessToken;
-    }
-    if (authToken) {
-      serverHeaders['x_amzn_oidc_accesstoken'] = authToken;
-      serverHeaders['Authorization'] = `Bearer ${authToken}`;
-    }
-    try {
-      const client = new Client({
-          name: mcpServer.name,
-          version: "1.0.0"
-        });
-      try {
-        const transport: StreamableHTTPClientTransport = new StreamableHTTPClientTransport(new URL(mcpServer.url), {
-          eventSourceInit: {
-            fetch: (input, init) =>
-              fetch(input, {
-                ...init,
-                headers: serverHeaders
-              }),
-          },
-          requestInit: {
-            headers: serverHeaders
-          }
-        });
-        await client.connect(transport);
-        console.log("Connected using Streamable HTTP transport");
-      } catch (error) {
-        console.log("Streamable HTTP connection failed, falling back to SSE transport");
-        const sseTransport = new SSEClientTransport(new URL(mcpServer.url), {
-          eventSourceInit: {
-            fetch: (input, init) =>
-              fetch(input, {
-                ...init,
-                headers: serverHeaders
-              }),
-          },
-          requestInit: {
-            headers: serverHeaders
-          }
-        });
-        await client.connect(sseTransport);
-        console.log("Connected using SSE transport");
-      }
-
-      const mcpTool = await loadMcpTools(mcpServer.url, client);
-
-      mcpTools.push(...mcpTool);
-    } catch (error) {
-      console.log(`Error trying to access tool: ${mcpServer.name}`, error);
-    }
-  }
+  let mcpTools = await getTools(selectedMcpServers, authToken);
+  console.log(mcpTools);
 
   let agent = createReactAgent({
     llm: agentModel,
