+++ frontend/src/logic/get-tools.ts
@@ -1,0 +1,71 @@
+import { Client } from '@modelcontextprotocol/sdk/client/index.js';
+import { SSEClientTransport } from '@modelcontextprotocol/sdk/client/sse.js';
+import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
+import { loadMcpTools } from '@langchain/mcp-adapters';
+import type { MCP_SERVER } from './get-servers';
+
+
+export const getTools = async (servers: MCP_SERVER[], authToken: string ) => {
+
+  let mcpTools = [];
+
+  for (const mcpServer of servers) {
+    const serverHeaders: any = {};
+    if (mcpServer.accessToken) {
+      serverHeaders['x-external-access-token'] = mcpServer.accessToken;
+    }
+    if (authToken) {
+      serverHeaders['x_amzn_oidc_accesstoken'] = authToken;
+      serverHeaders['Authorization'] = `Bearer ${authToken}`;
+    }
+    try {
+      const client = new Client({
+          name: mcpServer.name,
+          version: "1.0.0"
+        });
+      try {
+        const transport: StreamableHTTPClientTransport = new StreamableHTTPClientTransport(new URL(mcpServer.url), {
+          eventSourceInit: {
+            fetch: (input, init) =>
+              fetch(input, {
+                ...init,
+                headers: serverHeaders
+              }),
+          },
+          requestInit: {
+            headers: serverHeaders
+          }
+        });
+        await client.connect(transport);
+        console.log("Connected using Streamable HTTP transport");
+      } catch (error) {
+        console.log("Streamable HTTP connection failed, falling back to SSE transport");
+        const sseTransport = new SSEClientTransport(new URL(mcpServer.url), {
+          eventSourceInit: {
+            fetch: (input, init) =>
+              fetch(input, {
+                ...init,
+                headers: serverHeaders
+              }),
+          },
+          requestInit: {
+            headers: serverHeaders
+          }
+        });
+        await client.connect(sseTransport);
+        console.log("Connected using SSE transport");
+      }
+
+      const mcpTool = await loadMcpTools(mcpServer.url, client);
+      console.log(mcpTool);
+
+      mcpTools.push(...mcpTool);
+      mcpTools[mcpTools.length - 1].server = mcpServer;
+    } catch (error) {
+      console.log(`Error trying to access tool: ${mcpServer.name}`, error);
+    }
+  }
+
+  return mcpTools;
+
+};
